pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "openclaw-security"
version = "0.7.0"
description = "AI Agent Security Assessment and Governance Toolkit"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "OpenClaw Security Team"},
]
keywords = ["security", "ai", "agent", "llm", "mcp", "assessment", "governance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: Security",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "pyyaml>=6.0",
    "cryptography>=41.0.0",
    "watchdog>=3.0.0",
    "aiohttp>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "flake8>=6.0.0",
    "pylint>=2.17.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
openclaw-scan = "security_detector:main"
openclaw-harden = "security_hardening:main"
openclaw-multi-scan = "multi_platform_scanner:main"

[project.urls]
Homepage = "https://github.com/openclaw-security/agent-security-guide"
Documentation = "https://github.com/openclaw-security/agent-security-guide/wiki"
Repository = "https://github.com/openclaw-security/agent-security-guide"
Issues = "https://github.com/openclaw-security/agent-security-guide/issues"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
